+++ android/engine/src/test/java/org/smartregister/fhircore/engine/ui/appsetting/AppSettingViewModelTest.kt
@@ -47,13 +47,6 @@ class AppSettingViewModelTest : RobolectricTest() {
   }
 
   @Test
-  fun testOnRememberAppChecked() {
-    appSettingViewModel.onRememberAppChecked(true)
-    Assert.assertNotNull(appSettingViewModel.rememberApp.value)
-    Assert.assertEquals(true, appSettingViewModel.rememberApp.value)
-  }
-
-  @Test
   fun testLoadConfigurations() = runBlockingTest {
     coEvery { appSettingViewModel.fhirResourceDataSource.loadData(any()) } returns
       Bundle().apply { addEntry().resource = Composition() }
@@ -79,5 +72,23 @@ class AppSettingViewModelTest : RobolectricTest() {
 
     coVerify { appSettingViewModel.fhirResourceDataSource.loadData(any()) }
     coVerify { appSettingViewModel.defaultRepository.save(any()) }
+  }
+
+  @Test
+  fun testHasDebugSuffix_withSuffix_shouldReturn_true() {
+    appSettingViewModel.appId.value = "default/debug"
+    Assert.assertTrue(appSettingViewModel.hasDebugSuffix()!!)
+  }
+
+  @Test
+  fun testHasDebugSuffix_noSuffix_shouldReturn_false() {
+    appSettingViewModel.appId.value = "default"
+    Assert.assertFalse(appSettingViewModel.hasDebugSuffix()!!)
+  }
+
+  @Test
+  fun testHasDebugSuffix_emptyAppId_shouldReturn_null() {
+    appSettingViewModel.appId.value = ""
+    Assert.assertNull(appSettingViewModel.hasDebugSuffix())
   }
 }